Added bower as a task in the gulpfile and changed the postinstall npm script to run it.
This means that when you run the `nnpm install` step the postinstall fires of the gulp bower
step and gets the test dependencies.

We don't need `npm install´ to build the files every time because we commit the files. We
commit the files because this means that they are easier to use by projects that has this
project as a bower dependency. (They don't need to minify the files before use)