Added a mocked version of the rlog service
If you include this file in your tests, either instead of or overriding
the non mocked service it will stop calling the $http service. To
override the service you only need to add this file after you add the
original service definition.

The reason you might want to use the mocked version of this service is
that it is bad practice to do network calls from your unit tests, and
that you don't need to test this service in your tests, it has its own
tests.